feat(better_player): add onSetResolution callback to BetterPlayerDataSource
This commit adds a new `onSetResolution` callback to the
`BetterPlayerDataSource` class. This callback is called
whenever the video resolution is changed.  This allows for
custom actions to be performed when the resolution changes,
such as updating UI elements or performing other related
tasks.  The callback is optional and can be omitted if not
needed.  It's also added to the `_buildResolutionSelectionRow`
function to call the callback after setting the resolution.